% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_rp.R
\name{adjust_rp}
\alias{adjust_rp}
\title{Adjusts experimentally elicited risk preferences using the methodology of Turner and Landry, 2021}
\usage{
adjust_rp(
  mc_reps,
  large_adjustment,
  small_adjustment,
  lottery_probs_1,
  lottery_probs_2,
  lottery_payoffs_1,
  lottery_payoffs_2,
  rp_lb,
  rp_ub,
  initial_wealth,
  sub_beliefs,
  utility_function,
  lottery_choice,
  returned_obj = "midpoint",
  rp_resolution = 0.01
)
}
\arguments{
\item{mc_reps}{numerical value representing the number of monte-carlo replications to use.}

\item{large_adjustment}{numerical value representing the upper bound of the large adjustment as described by Turner and Landry, 2021}

\item{small_adjustment}{numerical value representing the upper bound of the small adjustment as described by Turner and Landry, 2021}

\item{lottery_probs_1}{a numerical vector representing the probabilities of the first outcome occurring in each lottery}

\item{lottery_probs_2}{a numerical vector representing the probabilities of the second outcome occurring in each lottery}

\item{lottery_payoffs_1}{a numerical vector representing the payoffs if the first outcome occurs in each lottery}

\item{lottery_payoffs_2}{a numerical vector representing the payoffs if the second outcome occurs in each lottery}

\item{rp_lb}{a numerical value representing the upper bound to consider for the risk preference coefficient.}

\item{rp_ub}{a numerical value representing the lower bound to consider for the risk preference coefficient.}

\item{initial_wealth}{a numerical value representing the initial wealth to use for each respondent}

\item{sub_beliefs}{a vector representing likert scale responses to each of the subjective probability belief elicitation questions in Turner and Landry, 2021}

\item{utility_function}{a character vector representing the utility function to use for deriving the risk prefrences. The only current option is "crra".}

\item{lottery_choice}{a numeric value representing the observed lottery choice of the survey respondent}

\item{returned_obj}{a character vector representing how adjusted risk preferences should be returned.
Option are "range" or "midpoint". "range" will return two values representing the lower and upper bounds on the risk preference
coefficeients that are compatiable with the respondent's observed choices and survey responses. "midpoint" will return a single
value that is the midpoing of the upper and lower bounds.}

\item{rp_resolution}{a numeric value representing the resolution to use for identifying risk preferences.
For example, a value of .01, will search the risk preference parameter space in 0.01 increments. This value can
be adjusted to speed up computation if fine resolution is not required.}
}
\value{
returns either a range or midpoint (depending on the value of returned_obj) representing the adjusted risk preferences
}
\description{
Adjusts experimentally elicited risk preferences using the methodology of Turner and Landry, 2021
}
\examples{
adjust_rp(mc_reps = 100, large_adjustment = .10, small_adjustment = .05,
rp_lb = -2,rp_ub = 2,rp_resolution = .01,lottery_probs_1 = c(1,.5,.225,.125,.025),
lottery_probs_2 = c(0,.5,.775,.875,.975),lottery_payoffs_1 = c(5,8,22,60,325),
lottery_payoffs_2 = c(0,3,2,0,0),sub_beliefs = c(1,2,3,4),lottery_choice = 3,
utility_function = "crra",initial_wealth = .0000001,returned_obj = "midpoint")

}
