% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{crra_u}
\alias{crra_u}
\title{Returns CRRA utility}
\usage{
crra_u(x, r)
}
\arguments{
\item{x}{numeric value representing wealth}

\item{r}{numeric value representing the coefficient of relative risk aversion}
}
\value{
Returns a numeric value representing the level of utility
}
\description{
Returns CRRA utility
}
\examples{
crra_u(100,.2)
}
